Minor changes
I've removed the article height error, added in max-height for the images to try and align them better.

Made minor changes to the homepage, to bring the desciptions to a similar length to help with alignment.

Unfortunately, because I'm running my test site on a different sever to the main site, I'm unable to use a lot of the php code and related images. I'm having to make changes to the live site using chrome, then make the changes to my files, then commit them. It makes testing very difficult. Any suggestions on how I can get around this?